failed
I think I need to just give in and use regular java to do the thing. There's no real reason to force it to be javascript. I know how it works a bit more now so I can try to make something else with it if I'm that looking to have something done..